Refactored Expression extensions
Moved expression extesnion methods into a single file and refactored some of the log reviewing code
