+++ FulcrumInjector/AppSettings.json
@@ -335,7 +335,7 @@
       ]
     },
     {
-      "SettingSectionTitle": "PassThru Regex Settings",
+      "SettingSectionTitle": "PassThru DLL Output Regex Settings",
       "SettingsEntries": [
         {
           "SettingName": "PTOpen Regex",
@@ -466,7 +466,7 @@
       ]
     },
     {
-      "SettingSectionTitle": "PassThru Syntax Settings",
+      "SettingSectionTitle": "PassThru DLL Output Syntax Settings",
       "SettingsEntries": [
         {
           "SettingName": "Type And Time Colors",
