+++ FulcrumInjector/FulcrumLogic/ExtensionClasses/TextTableGenerator.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace FulcrumInjector.FulcrumLogic
+namespace FulcrumInjector.FulcrumLogic.ExtensionClasses
 {
     /// <summary>
     /// Class used to convert a set of tuples of values into an ASCII printed text table.
