+++ FulcrumInjector/FulcrumViewContent/Models/SettingsModels/FulcrumSettingsShare.cs
@@ -48,13 +48,13 @@ namespace FulcrumInjector.FulcrumViewContent.Models.SettingsModels
         // Settings for Regex Objects during parsing.
         public static SettingsEntryCollectionModel InjectorRegexSettings =>
             SettingsEntrySets?.FirstOrDefault(SettingObj =>
-                SettingObj.SettingSectionTitle.Contains("PassThru Regex Settings")
+                SettingObj.SettingSectionTitle.Contains("PassThru DLL Output Regex Settings")
             ) ?? new SettingsEntryCollectionModel("PassThru Regex Settings", Array.Empty<SettingsEntryModel>());
 
         // Settings for color output during formatting for PT Output
         public static SettingsEntryCollectionModel InjectorDllSyntaxSettings =>
             SettingsEntrySets?.FirstOrDefault(SettingObj =>
-                SettingObj.SettingSectionTitle.Contains("PassThru Syntax Settings")
+                SettingObj.SettingSectionTitle.Contains("PassThru DLL Output Syntax Settings")
             ) ?? new SettingsEntryCollectionModel("PassThru Syntax Settings", Array.Empty<SettingsEntryModel>());
 
         // TODO: FINISH DEBUG SYNTAX SETTINGS HERE
